+++ packages/nuxt/src/runtime/composables.ts
@@ -1,0 +1,4 @@
+import { useNuxtApp } from '#app'
+export * from 'pinia'
+
+export const usePinia = () => useNuxtApp().$pinia