+++ packages/pinia/__tests__/vitest-setup.ts
@@ -1,0 +1,6 @@
+import { beforeEach } from 'vitest'
+import { setActivePinia } from '../src'
+
+beforeEach(() => {
+  setActivePinia(undefined)
+})